refact(contexts): separate contexts states for inbounds, nodes, hosts
- Renaming files from tsx to ts
- Relocating types to @/types
- Separate state contexts for inbounds, nodes, hosts
- Moving QRCode and Subscriptions link states to user
