+++ app/dashboard/src/contexts/CoreSettingsContext.ts
