+++ app/dashboard/src/types/Protocol.ts
@@ -1,0 +1,3 @@
+
+export type ProtocolType = 'vmess' | 'vless' | 'trojan' | 'shadowsocks';
+
